Add a hook to pick which validator to use
