+++ src/schema/validator.js
@@ -1,15 +1,6 @@
-import ajv from 'ajv';
-import { isRelativeURL, isValidVersionString } from './formats';
+import firefoxValidate from './firefox-validator';
+import linterValidate from './linter-validator';
 
-export var schemaObject = require('json!schema/manifest-schema');
-
-var validator = ajv({
-  allErrors: true,
-  errorDataPath: 'property',
-  jsonPointers: true,
-  verbose: true,
-});
-validator.addFormat('versionString', isValidVersionString);
-validator.addFormat('relativeURL', isRelativeURL);
-
-export default validator.compile(schemaObject);
+export default process.env.USE_FIREFOX_SCHEMAS
+  ? firefoxValidate
+  : linterValidate;
